+++ blur-cmp/src/iosMain/kotlin/io/github/ezoushen/blur/cmp/RegisterBackdropCaptureSource.ios.kt
@@ -1,0 +1,6 @@
+package io.github.ezoushen.blur.cmp
+
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun RegisterBackdropCaptureSource() = Unit